🚚(frontend) move current user store
Move currentUserStore from lti site to lib-components.
