+++ src/frontend/packages/lib_components/src/hooks/stores/index.ts
@@ -1,0 +1,1 @@
+export * from './useCurrentUser';
