+++ src/frontend/packages/lib_components/src/index.ts
@@ -1,2 +1,3 @@
 export * from './common';
+export * from './hooks';
 export * from './types';
